+++ Sources/RichTextKit/Export/RichTextExportUrlResolver+FileManager.swift
@@ -8,24 +8,20 @@
 
 import Foundation
 
-/**
- This is a typealias for the `FileManager` class, since it's
- the standard way to resolve export file urls.
- */
+/// This is a typealias for the `FileManager` since it's the
+/// standard way to resolve export file urls.
 public typealias StandardRichTextExportUrlResolver = FileManager
 
 extension FileManager: RichTextExportUrlResolver {}
 
 public extension FileManager {
 
-    /**
-     Try to generate a file url in a certain directory.
-
-     - Parameters:
-       - fileName: The preferred file name.
-       - extensions: The file extension.
-       - directory: The directory in which to generate an url.
-     */
+    /// Try to generate a file url in a certain directory.
+    ///
+    /// - Parameters:
+    ///   - fileName: The preferred file name.
+    ///   - extension: The file extension.
+    ///   - directory: The directory in which to generate an url.
     func fileUrl(
         withName fileName: String,
         extension: String,
@@ -39,19 +35,17 @@ public extension FileManager {
         return fileUrl
     }
 
-    /**
-     Try to generate a unique file url in a certain directory.
-
-     If needed, the function appends a counter until the url
-     is unique. This means that the resulting url for a file
-     url that has the file name `myFile.txt` may result in a
-     url that has the file name `myFile-1.txt`.
-
-     - Parameters:
-       - fileName: The preferred file name.
-       - extensions: The file extension.
-       - directory: The directory in which to generate an url.
-     */
+    /// Try to generate a file url in a certain directory.
+    ///
+    /// If needed, this function will append a counter until
+    /// the url is unique. This means that the resulting url
+    /// for a file with the name `myFile.txt` may instead be
+    /// a url with the file name `myFile-1.txt`.
+    ///
+    /// - Parameters:
+    ///   - fileName: The preferred file name.
+    ///   - extension: The file extension.
+    ///   - directory: The directory in which to generate an url.
     func uniqueFileUrl(
         withName fileName: String,
         extension: String,
@@ -62,18 +56,16 @@ public extension FileManager {
         return uniqueUrl
     }
 
-    /**
-     Get a unique url for the provided `url`, to ensure that
-     no existing folder or file exists there.
-
-     If needed, the function appends a counter until the url
-     is unique. This means that the resulting url for a file
-     url that has the file name `myFile.txt` may result in a
-     url that has the file name `myFile-1.txt`.
-
-     - Parameters:
-       - url: The url to generate a unique url for.
-     */
+    /// Get a unique url for the provided `url` to make sure
+    /// that no existing folder or file exists there.
+    ///
+    /// If needed, this function will append a counter until
+    /// the url is unique. This means that the resulting url
+    /// for a file with the name `myFile.txt` may instead be
+    /// a url with the file name `myFile-1.txt`.
+    ///
+    /// - Parameters:
+    ///   - url: The url to generate a unique url for.
     func uniqueUrl(for url: URL) -> URL {
         if !fileExists(atPath: url.path) { return url }
         let fileExtension = url.pathExtension