+++ Sources/RichTextKit/_Deprecated/NSMutableParagraphStyle+Custom.swift
@@ -14,6 +14,7 @@ import UIKit
 import AppKit
 #endif
 
+@available(*, deprecated, message: "This is no longer used within the library and will be removed in the next major version bump.")
 extension NSMutableParagraphStyle {
 
     convenience init(