+++ RELEASE_NOTES.md
@@ -8,11 +8,11 @@ Until then, minor updates may remove deprecated features and introduce breaking 
 
 ## 1.2
 
-This version adds key path-based paragraph functions to the `RichTextViewComponent`. This means that you can now get and set the following values with a single set of functions:
+This version adds key path-based paragraph functions to the `RichTextViewComponent`. This means that you can now get and set the following values with a single set of functions, pickers, steppers and toggles:
 
 `alignment`, `allowsDefaultTighteningForTruncation`, `baseWritingDirection`, `defaultTabInterval`, `firstLineHeadIndent`, `headIndent`, `hyphenationFactor`, `lineBreakMode`, `lineBreakStrategy`, `lineHeightMultiple`, `lineSpacing`, `maximumLineHeight`, `minimumLineHeight`, `paragraphSpacing`, `paragraphSpacingBefore`, `tabStops`, `tailIndent`, `usesDefaultHyphenation`.
 
-As a result, all parts of the library that handled individual value types have been deprecated. This makes the library tighter, and also a lot more capable, with less code. 
+As a result, all parts of the library that handled individual value types have been deprecated. This makes the library tighter, and a lot more capable with less code. 
 
 ### ✨ Features
 
@@ -37,8 +37,8 @@ As a result, all parts of the library that handled individual value types have b
 
 ### 🗑️ Deprecations 
 
+* `RichTextAlignment` has been deprecated due to the new paragraph-based features.
 * `RichTextAlignment.Picker` has been deprecated due to the new `Picker` capabilities.
-* `RichTextAttributeReader` has deprecated a lot of functions.